p2p protocol over Web RTC
In order to bring the bitcoin p2p protocol to web apps, we can leverage
Web RTC, which allows p2p data connections. The catch is that is not not
possible to connect directly from a browser to the real bitcoin p2p
network, because the real bitcoin p2p network does not already support
Web RTC. Thus we will need to create gateways that link the Web RTC
network to the normal network.

I have made a first step towards this by making the Connection and
Server classes work over Web RTC. You can now connect another peer if
you share the same rendezvous server, and then share bitcoin p2p
messages. The only messages I have correctly implemented so far are the
ping and pong messages. You can ping, and that person will pong back.

I have used PeerJS to provide the framework for using Web RTC. PeerJS
hides a huge amount of complexity. I have erred away from adding
dependencies to fullnode to keep it easier to audit, but in this case I
make an exception, because it would take me too long to implement a
simplistic version of what PeerJS does. It is worth the extra
dependency.

Note that there are two components to PeerJS. There is the part that
runs in a web browser for connecting to other peers, and there is the
server component, which facilitates rendezvous of peers. Two peers must
use the same rendezvous server to find each other. As far as I can tell,
it is possible to use multiple rendezvous servers at the same time. Thus
one way to bridge the Web RTC network to the normal network is for many
people to run rendezvous servers. Anybody who runs a "full" full node
also runs a PeerJS rendezvous server, allowing web-only clients to do
peer discovery on their server.

Also note that in order to run the tests in the browser, the rendezvous
server must be running. I have set it up so that the testapp runs the
rendezvous server and so does karma.