add network abstractions
In order to get rid of the checks in "if process.browser" everywhere
inside Server and Connection, I have added new abstraction layers
Network and Netbufs. Network is an abstraction for either the normal
internet, or a PeerJS network around a rendezvous server. You can
connect to other peers on the same network. Netbufs is an abstraction
for a data channel on a network. Perhaps "channel" would have been a
better word. For now, I am simply adding them to the library, however
the next step is to use them inside Server and Connection in order to
get rid of all the "if process.browser" checks.