fix things for "node --harmony"
fullnode is written in ES6 and is intended for use in io.js, or whatever
the next major version of node.js is that supports ES6 by default.
However, node 0.12 supports ES6 features with the --harmony flag. I have
made two minor fixes to make the library work in node 0.12 harmony. One
was to fix an accidental use of the new simpler object literal notation
that doesn't seem supported in 0.12. Another was to fix a scoping issue
in fromBitcoindString that I suspect is a bug in the V8 version used by
node 0.12. With these two fixes, you can now use node 0.12, and run
"mocha --harmony" to run the tests, which all pass. Unfortunately things
don't work with node 0.10, because it doesn't support enough features of
ES6, even with --harmony.