await -> wait, and more semistandard --format
'await' is actually a reserved keyword for use in a future version of
javascript. The next version of javacript will be ES7 (or maybe ES2016?) and
will likely use the 'await' keyword to mean something different than what we
are using it for here. (Albeit with a similar asynchronous intent). This was
creating problems with semistandard, which was erroring out on this and not
formatting some files. I have fixed this by renaming 'await' to 'wait', and
then running semistandard on the remaining files.

semistandard --format is not able to fix all non-standard style issues. I have
yet to bother to fix any of those by hand, but I will do so over time.