es6ify -> babelify
It seems babel has become the standard for transpiling javascript, so I've
switched to babel. There is only one gotcha with babel as opposed to traceur
(which is used by es6ify), which is that you need to include the polyfill on a
page exactly one time, and not more. Thus I've put it in the fullnode bundle,
and nowhere else. For instance, it is not in the tests - only in the main
bundle. The tests now include the bundle once to be sure the polyfill is being
loaded exactly once. It seems workers must load the polyfill exactly once
themselves. Since they also load the bundle, this works as expected.