add some test vectors from bitcoinjs-lib
I was debugging a problem related to compact signatures in Datt that I thought
might be from fullnode, so I tested against the relevant bitcoinjs-lib test
vectors. It turns out the problem was actually in Datt, however it's still nice
to keep these test vectors. bitcoinjs-lib has an amazing suite of test vectors
and it would be worthwhile to test against all of them. I have also done this
in the past as can be seen in the commit history. This time, I have only
imported a tiny subset, but more should be added.