update bn.js and fix test issue
There appears to have been a bug in the tests for umod. "umod" means unsigned
modulus, or modulus performed on numbers that are assumed to be unsigned. And
yet the test was plugging in signed numbers. There should be no reason to test
against signed numbers in the umod tests, since this function should never be
used on signed numbers. Fedor's code does not check if the numbers are signed
and does not throw an error if you pass it signed numbers - probably because he
didn't want to have an extra conditional in there which would slow things down.