add "bitcoin" back to README
I had previously removed references to "bitcoin" to promote the notion that the
more important property was the blockchain, and to get away from the negative
connotations that "bitcoin" has in some circles. However, I think it is
important to mention the word "bitcoin" so that people who bother to read the
README know that this isn't some other blockchain.