asyncMethod -> asyncObjectMethod
Right now the only thing you can do with the works is give them a fullnode
operation, and perform a method on them, with a list of arguments that can only
contain basic types (i.e., numbers and strings, but not buffers or other
fullnode objects). I will need more powerful behavior than this. We will need
to be able to perform class methods rather than object methods, i.e.
Hash.sha256, and we will also need to be able to pass fullnode objects and
Buffers in the list of arguments. The first step towards this goal is just to
rename asyncMethod to asyncObjectMethod to indicate that this is a way to
perform an operation on an object, not a class. Later we will also need to add
support for operating on classes as well as passing more sophisticated types in
the arguments.