support sending complex arguments to workers
It was the case that you could only send standard javascript types in the
arguments to the workers, but this was insufficient, because you might want to
send buffers or fullnode objects in the arguments. I have now added support for
sending complex arguments, and also some tests by using this functionality for
public keys. You can now derive a public key from a private key using this
strategy. I needed to add support for to/fromFastBuffer for pubkeys for this to
work, in particular zero sized, i.e. blank, public keys.