largest tx test vector
There is a very large transaction on the blockchain whose size is about 1 mb.
I've added a test to be sure we can hash this particular transaction correctly.
In an upcoming commit, I will also test that we can hash this transaction
correctly in a worker. The reason for this test is to be sure we can send a lot
of data to a worker without breaking things.

The txid is: bb41a757f405890fb0f5856228e23b715702d714d59bf2b1feb70d8b2b4e3e08