allow workers to handle large data
In node with child process forks, it turns out that the easiest way to send
data in pieces to a worker is to let node split up the data for us. It will
automatically break up a 1 mb piece of data into small pieces, and send them
one at a time. We just have to be able to assemble the pieces inside the
worker. This now works. In a browser, it is slightly different - evidently it
is possible to send a 1 mb piece of data in one go. However, in the interest of
keeping the worker code equivalent between node and the browser, I have kept
the same assembly logic in both places.