fix hanging connections for proper process exit
The workers and network connections were staying alive at the end of the tests,
so the process had to be manually exited with "process.exit()". That was not
elegant - the right solution is to properly close the workers and network
connections at the end of the process so that the tests end correctly.