derive merkle from list of buffers
You can now pass in a list of buffers, such as a list of transactions, and
derive the merkle root. Note that I have not yet tested this against bitcoin.
Although deriving the merkle root works, it may not derive it in a
byte-for-byte compatible way with bitcoin. Further testing may require changes.