Msg
The Msg class was already written in a previous commit of fullnode. It was
there in v0.12.0 but was removed by 0.12.1. It was removed because the
networking code needed to be rearchitected. However, the Msg class is mostly
acceptable, so I have added it back in, with two differences: .fromBuffers is
now .genFromBuffers, and Msg is now intended to use inheritance for other Msg
types.

First, what was the .fromBuffers method will now be called .genFromBuffers. The
reason for this is to clarify that the method is a generator and thus needs to
be used as such, and also to dinstinguish it from the other .fromBuffers method
used in Merkle, which has a different meaning (creating an object from an array
of buffers, rather than from a generator). I have also restored the
.genFromBuffers method in Struct (which used to be called .fromBuffers), which
simply throws an error an indicates that you need to implement it if you
inherit from Struct.

Second, the new version does not have the ".fromPing" or ".toPing" and so on.
The way we will distinguish between different messages is to use inheritance,
which is probably the way almost all programmers would solve this problem. The
path we were going down before would have put way too much logic in this one
class. Instead, only "msg" logic will be here, and a separate class will be
created for "msgping" and for all other message types. The reason to use
inheritance comes from my experience NOT using inheritance in the Datt p2p
networking, and realizing how painful it was that a "msgping" was not also a
"msg".