Msg.prototype.asyncSetData
When creating a new message, you will want to set the data (a.k.a. "payload"),
and setting the data requires computing a checksum, which is a hash function,
and is therefore blocking on the main thread. It is better to send the data to
a worker and compute the hash there so as to not block the main thread. So I
have implemented, first, an asyncChecksum method, which does exactly that, and,
secondly, an asyncSetData method, which sets the data and uses the
asyncChecksum method to actually find the checksum.