move signature fill logic to separate methods
Right now, only the synchronous sign methods works on p2sh multisig. We need to
add support for asynchronous signing. To do that we need to put most of the
logic into reusable methods, including the signature fill code (the stuff that
actually places the signature into the appropriate place in a transaction).