restore npm-shrinkwrap.json
A couple of versions ago, I encountered an issue with npm shrinkwrap and
shrinkpack. The shrinkwrap file was referencing files that didn't exist. The
problem had something to do with cached packages referencing files in the
node_shrinkwrap that didn't exist. This was caused by shrinkpack. The problem
is solved by clearing cache and running "npm shrinkwrap" again.