OP_CHECKLOCKTIMEVERIFY (CLTV)
A new operation was added to Bitcoin Core recently via a softfork.
OP_CHECKLOCKTIMEVERIFY, or CHECKLOCKTIMEVERIFY, or CLTV, validates that the
current transaction's locktime is higher than some value. This makes it
possible to lock funds on the blockchain until a certain date. If you use CLTV,
you cannot create a spend transaction until the locktime has passed. Before
CLTV, there was no way to lock funds on the blockchain until a certain date -
locktime only prevented a transaction from being valid until a future date,
which means that transaction was not actually on the blockchain. But now a
transaction can be on the blockchain and unspendable until a future date.

Notes on the changes:
- The code for CLTV comes from Bitcoin Core v0.12.0, commit
  188ca9c305d3dd0fb462b9d6a44048b1d99a05f3, and has been translated to
  javascript while keeping it as similar as possible to the original code,
  including comments.
- BN.prototype.fromScriptNumBuffer had to be modified to support reading in a
  5-byte number.
- Interp needed nin added to the constructor, as it is one of the properties of
  the interpreter. It always should have been in the constructor - leaving it
  out was a mistake.
- Interp.LOCKTIME_THRESHOLD has been added. This constant existed before CLTV
  and is used to implement a full node.
- Update script and tx tests have been copied from Bitcoin Core and pass. Those
  tests are script_valid, script_invalid, tx_valid, and tx_invalid.
- Bitcoin Core has been modified to allow a signature to have 0 length,
  providing a convenient mechanism to support invalid signatures for testing
  purposes, so those changes are reflected here.