MsgBlock
I'm going to begin adding all bitcoin p2p network messages, or at least those
necessary to download and maintain the blockchain. I'm literally going in
alphabetical order, so "block" is first. A "block" message, denoted MsgBlock in
Fullnode, simply transmits a block. The validation of a message block confirms
whether it is a valid "block" message, but not whether the contained block is
actually valid.