allow sorting public keys, and sort by default
It is "standard", at least in some circles, to sort the public keys in a
redeemScript. The reason for this is because the p2sh address formed from the
public keys is sensitive to the order, and in order to gurantee reproducibility
of the address, the simplest answer is to sort them.

An old thread on the mailing list about this:

https://lists.linuxfoundation.org/pipermail/bitcoin-dev/2014-March/004629.html

The code for sorting was taken from my unused project "treasure", which was for
managing keys in a multisig wallet:

https://github.com/ryanxcharles/treasure

treasure preceded what would become Copay.