OP_CHECKSEQUENCEVERIFY (CSV)
Bitcoin Core v0.12.1 was released a few days ago. It included three soft forks,
one of which was interpreting the sequence number in an input to be a relative
locktime, and another being a new opcode CHECKSEQUENCEVERIFY (CSV) which checks
that a sequence number was high enough. I have copied the code for CSV almost
exactly from Bitcoin Core, including comments, with as few changes as possible.
They did not seem to have updated the script_valid.json or script_invalid.json
tests, but they did update the tx_valid.json and tx_invalid.json tests, which
pass.

Note that there are some subtleties with respect to the way the bitwise
operations are working in the new checkSequence field. seqnum is actually an
unsigned integer, but javascript bitwise operators operator on signed integers.
However, javascript converts unsigned integers into their two's complement form
when operating on them with bitwise operators, which is exactly what we want -
we are operating on the "binary" form. The end result is that we can use the
same code from Bitcoin Core with no modifications, except to note that there
are some subtle conversions going on when an integer is over the max size for a
signed integer. I have not tried to test the edge cases - if Bitcoin Core's
tests do not test these edge cases, then it is possible there is a bug here.