fix js standard nits
The latest version of standard made a bunch of changes to what it regards as
standard javascript formatting, including requiring a space after the * in a
generator, and requiring camel case for all variables (even though in some
cases I stuck to non-camel case specifically to keep the code as similar to
standards as possible). I didn't notice this until now. All nits are now fixed.