class Point
Point is one of the most unusual classes in Fullnode because it is extended
from an external library, Fedor Indutny's elliptic curve library. His code was
not easy to extend, and that is why there is some javascript weirdness with the
way it is extended. Fortunately, converting it to an ES2015 class seems to
"just work" in spite of the weirdness.