version -> versionBytesNum or versionByteNum
One of the p2p messages, "version", uses a word, "version", which we already
use throughout Fullnode. I think the best way to handle this is to rename all
instances of "version" that we currently have, which all reference a number, to
"versionBytesNum" if it is multiple bytes or "versionByteNum" if it is only one
byte. This way it is clear whether we are talking about a "versionByte" or
"versionBytes", which are bytes, or a "version", which is a specific data
structure used in the p2p protocol.