use default function arguments in most places
The whole reason the "initialize" pattern was originally created in Yours
Bitcoin was because of the lack of default function arguments in javascript.
The latest version of Javascript, ES2015, and associated update to Node 6.1,
allows for default function arguments. I have updated most constructors to use
default function arguments, with the notable exception of all the Msg classes
since they use the initialize function is a slightly more sophisticated way.