experiments show 6 processes is fastest
On my machine with 4 cores and 8 (hyperthreaded) virtual cores, it seems 6
processes in the sweet spot for these tests. This may be a fluke on this test
suite, but I will run similar tests on the other Yours repositories later to
see if similar results hold. Here are the results of my timing tests:

Time tests, "real" is the actual time it took, "user" is CPU time:

- real (user)

1 process
- 8.090s (6.476s)
- 7.361s (5.924s)
- 7.364s (6.116s)

4 processes:

- 3.821s (9.620s)
- 3.890s (9.724s)
- 3.893s (3.893s)

6 processes:

- 3.656s (13.476s)
- 3.620s (13.432s)
- 3.453s (13.008s)

8 processes:

- 3.630s (15.456s)
- 3.531s (15.348s)
- 3.798s (15.860s)

16 processes:

- 4.579s (21.324s)
- 4.634s (21.464s)
- 4.607s (21.408s)

I also ran separate tests to directly compare 6 processes to 8 processes
and 6 processes was consistenty faster, albeit only by a small margin.